LayoutTests/http/tests/websocket/tests/hybi/bufferedAmount-after-close-in-busy-expected.txt

Web Socket bufferedAmount after closed in busy

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS bufferedAmountBeforeClose + closeFrameSize >= bufferedAmountAfterClose is true
Closed.
PASS ws.readyState is 3
PASS ws.bufferedAmount <= bufferedAmountAfterClose is true
Testing send(string)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 27
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
Testing send(ArrayBuffer)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
Testing send(Blob)...
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 6
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 7
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 131
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 134
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65543
PASS ws.send(messageToSend) is false
PASS bufferedAmountDifference is 65550
PASS successfullyParsed is true

TEST COMPLETE
